Trying to revert application.yml to production working mode.
The site has crashed and it seems from the azure server log stream that there is a problem with the sql server dialect and JPA. I don't know why that is because it worked before. But this change will revert to exactly the same settings as it was when we ran the server on Azure in the beginning of August.